Fixed search legality issues around quiescence while in check, fixed mate-distance scoring and root ply handling, refreshed legacy NNUE accumulators correctly after 'position', corrected SFNNv14 forward-output scaling. Implemented the NNUE parity pass and the opening symptom of 'a2a3' is much healthier

+++ nnue/sfnnv14_threat_updates.odin
@@ -538,7 +538,7 @@ apply_threat_buffer_to_accumulator :: proc(
 
 		weight_offset := idx * SFNNV14_L1
 		for j in 0 ..< SFNNV14_L1 {
-			state.accumulation[perspective][j] -= sfnnv14_transformer.threat_weights[weight_offset + j]
+			state.accumulation[perspective][j] -= i16(sfnnv14_transformer.threat_weights[weight_offset + j])
 		}
 		psqt_offset := idx * SFNNV14_PSQT_BUCKETS
 		for j in 0 ..< SFNNV14_PSQT_BUCKETS {
@@ -554,7 +554,7 @@ apply_threat_buffer_to_accumulator :: proc(
 
 		weight_offset := idx * SFNNV14_L1
 		for j in 0 ..< SFNNV14_L1 {
-			state.accumulation[perspective][j] += sfnnv14_transformer.threat_weights[weight_offset + j]
+			state.accumulation[perspective][j] += i16(sfnnv14_transformer.threat_weights[weight_offset + j])
 		}
 		psqt_offset := idx * SFNNV14_PSQT_BUCKETS
 		for j in 0 ..< SFNNV14_PSQT_BUCKETS {
